+++ FitnessApp/Home/Models/Activity.swift
@@ -1,0 +1,19 @@
+//
+//  Activity.swift
+//  FitnessApp
+//
+//  Created by Mohamad Alayouni on 3/3/25.
+//
+
+import Foundation
+import SwiftUI
+
+struct Activity {
+    let id: Int
+    let title: String
+    let subtitle: String
+    let image: String
+    let tintColor: Color
+    let amount: String
+
+}
