+++ FitnessApp/Home/Models/Workout.swift
@@ -1,0 +1,20 @@
+//
+//  Workout.swift
+//  FitnessApp
+//
+//  Created by Mohamad Alayouni on 3/3/25.
+//
+
+import Foundation
+import SwiftUI
+
+struct Workout {
+    let id: Int
+    let title: String
+    let image: String
+    let tintColor: Color
+    let duration: String
+    let date: String
+    let calories: String
+    
+}
